requirements.txt
pyspark==3.3.0
tweepy==4.9.0
requests==2.28.1
pandas==1.4.4
retry==0.9.2
kafka-python==2.0.2